refactor: strictly map directory structure to DDD and Hexagonal Architecture
- Reorganize files into pure Bounded Contexts under `core/`.
- Shift unmapped configuration and schemas out of `spec/domains` and `core/` root.
- Move test suite structure to perfectly mirror the new physical `src/` constraints.
- Fix all cross-project import bounds to match the resulting directory migrations.
