+++ src/coreason_manifest/core/domains/mcp_contracts.py
