+++ tests/core/telemetry/test_telemetry.py
